Change i2 Query ALDB to Single Record Request
In testing I already saw a corrupted extended aldb record response.

It seemed too complicated to create a functional to verify the accuracy of the entire ALDB, and errors could still look like valid entries.

So instead, request a single entry at a time.

This is absolutely slower and puts more traffic on the network, but it is rather stable and robust.